Crate ex_client

Macros

line_error

Structs

ClientId

Client Id type used to identify a client.

Key

A key to the crypto box. Key is stored on the heap which makes it easier to erase. Makes use of the GuardedVec<u8> type to protect the data.

KeyStore

Provider

RecordHint

a record hint. Used as a hint to what this data is used for.

SecureBucket

Snapshot

Wrapper for the SnapshotState data structure.

SnapshotState

Data structure that is written to the snapshot.

Enums

EngineError

Error block for the Engine.

Error

Stronghold Client error block.

Location

A Location type used to specify where in the Stronghold a piece of data should be stored. A generic location specifies a non-versioned location while a counter location specifies a versioned location. The Counter location can be used to get the head of the version chain by passing in None as the counter index. Otherwise, counter records are referenced through their associated index. On Read, the None location is the latest record in the version chain while on Write, the None location is the next record in the version chain.

VaultError

Traits

LoadFromPath

A trait that allows a datatype to load and setup its internal data through the use of a path and some data.

Functions

derive_vault_id

Gets the VaultId from a specified path.

naive_kdf

public_key_inner

resolve_location

sr25519_sign_inner

Type Definitions

Result